- basic end-to-end for generating lip sync data automatically

+++ SCICompanionLib/Src/Util/PhonemeMap.h
@@ -5,7 +5,7 @@ class PhonemeMap
 public:
     PhonemeMap(const std::string &filename);
 
-    int PhonemeToCel(const std::string &phoneme);
+    uint16_t PhonemeToCel(const std::string &phoneme);
 
 private:
     std::string _filename;
